+++ tests/assets/multilanguage/pagination.tmpl.js
@@ -1,9 +1,17 @@
-export const lang = ["gl", "en"];
+// export const lang = ["gl", "en"];
 export const layout = "layout.njk";
 
-export default function* ({ search, paginate, mergeLanguages }) {
-  const gl = paginate(search.pages("lang=gl", "url=asc"));
-  const en = paginate(search.pages("lang=en", "url=asc"));
-
-  yield* mergeLanguages({ gl, en });
+export default function* ({ search, paginate }) {
+  for (const p of paginate(search.pages("lang=gl", "url=asc"))) {
+    yield {
+      ...p,
+      id: `page-${p.pagination.page}`,
+    };
+  }
+  for (const p of paginate(search.pages("lang=en", "url=asc"))) {
+    yield {
+      ...p,
+      id: `page-${p.pagination.page}`,
+    };
+  }
 }