+++ tests/utils.ts
@@ -4,7 +4,7 @@ import { basename, fromFileUrl, join } from "../deps/path.ts";
 import { printError } from "../core/errors.ts";
 import { DeepPartial } from "../core/utils.ts";
 
-import type { Page, Site, SiteOptions, SourceMap } from "../core.ts";
+import type { Data, Page, Site, SiteOptions, SourceMap } from "../core.ts";
 
 const cwUrl = import.meta.resolve("./");
 const cwd = fromFileUrl(import.meta.resolve("./"));
@@ -139,7 +139,7 @@ export async function assertSiteSnapshot(
     if (page.data.alternates) {
       // @ts-ignore: Remove alternates
       page.data.alternates = Object.keys(
-        page.data.alternates as Record<string, Page>,
+        page.data.alternates as Record<string, Data>,
       ).sort();
     }
     // Remote base path because it's different in the test environment